Fixed Heap with --debug-memory.
git-svn-id: svn://0x263a.de/tekuum_src@223 3ae3452d-76a7-4814-9b33-b51e13292e42

+++ idlib/Heap.cpp
@@ -1723,8 +1723,12 @@ void *Mem_Alloc16( const int size, const char *fileName, const int lineNumber ) 
 		return NULL;
 	}
 	void *mem = Mem_AllocDebugMemory( size, fileName, lineNumber, true );
+
 	// make sure the memory is 16 byte aligned
-	assert( ( ((int)mem) & 15) == 0 );
+	// Techyon RB: 64 bit fix, changed int to intptr_t
+	assert( ( ((intptr_t)mem) & 15) == 0 );
+	// Techyon END
+
 	return mem;
 }
 
@@ -1737,8 +1741,12 @@ void Mem_Free16( void *ptr, const char *fileName, const int lineNumber ) {
 	if ( !ptr ) {
 		return;
 	}
+
 	// make sure the memory is 16 byte aligned
-	assert( ( ((int)ptr) & 15) == 0 );
+	// Techyon RB: 64 bit fix, changed int to intptr_t
+	assert( ( ((intptr_t)ptr) & 15) == 0 );
+	// Techyon END
+
 	Mem_FreeDebugMemory( ptr, fileName, lineNumber, true );
 }
